CLEAN: move compiler-features.json generation to a separate file
